add interfaces to user api
