+++ packages/client/src/mocs/loginProps.ts
@@ -3,12 +3,8 @@ import { ILoginForm } from '../pages/login/login';
 const LoginProps: ILoginForm = {
   headerName: 'Вход в игру',
   linkTitle: 'создать аккаунт',
-  linkAction: () => {
-    console.log('router to registr page');
-  },
-  formAction: () => {
-    console.log('router to registr page');
-  },
+  linkAction: () => { console.log('router to login page') },
+  formAction: () => { console.log('router to submit form') },
   submitBtnName: 'Войти',
   inputsProps: [
     {
