+++ packages/client/src/utils/helpers.tsx
@@ -5,7 +5,7 @@ export const unAuthorizedRedirect = async () => {
   /*
   Когда подключим API авторизации здесь должен быть запрос пользователя.
   Если  не авторизован, то переход на страницу авторизации */
-  const user = null
+  const user = {}
   if (!user) {
     return redirect(links.login.path)
   }
@@ -16,7 +16,7 @@ export const authorizedRedirect = async () => {
   /*
   Когда подключим API авторизации здесь должен быть запрос пользователя.
   Если авторизован, то переход на страницу игры */
-  const user = null
+  const user = {}
   if (user) {
     return redirect(links.game.path)
   }
